Supply revision numbers for all builds
The packages will be built with REVISION in the environment. This may
not be enough to get the values linked in to the binaries where
necessary.

In addition, the values are propagated to the packages by the correct
filename and (for rpm) by setting `--iteration` and `--rpm-dist` on the
fpm invocation.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>