tests: Move img setup behind interfaces
This simplifies how all the things we need are created/accessed.

One big reason for this change is I was going to have to add a 3rd
return parameter for formatting package versions and thought things were
getting a bit out of hand.
The need for such formatting is that deb and rpm formating their
package names differently so we can't just use a generic formatter for
everything.
It is also greatly simpler (and faster) to encode the value
we expect instead of trying to read that from /etc/os-release.