tests: qemu: use virtual kernel
This one is available on both amd64 and arm64 so we don't need to branch
any platform branching (`-kvm` is not).
Also this gets rid of the `*` and need for installing kernel modules
separately. As it turns out, sometimes there are multiple versions of
the kernel available.

This change made it neccessary to require dhcp to be performed in the vm
rather than relying on the kernel command line since the neccessary
kernel configs aren't there.