pipeline: Conditionally fetch moby-tini
Because we do not publish directly to the mariner repo, moby-tini is not
available in the repo. As such, it has to be installed manually whenever
moby-tini is a dependency of the package being built. Currently, this is
only the case with moby-engine.

tldr; we need a workaround (for now) for mariner2/moby-engine.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>