pipeline: Fix globbing issue
An issue came up when the mariner folks started publishing both
`moby-containerd` and `moby-containerd-cc`. Because we were
indiscriminately globbing with `moby-containerd-<env_var>*`, if
`<env_var>` was empty, then the glob would be `moby-containerd-*`.

When parsed by dnf, this would catch `moby-containerd-cc` before
`moby-containerd`. Because those two packages have different versions
and conflict with each other, there were dependency resolution issues,
particularly when installing `moby-engine` because it depends on
`moby-containerd`.

This fix makes the globbing logic *slightly* more nuanced. In the future
it might be worth considering a more explicit specification of the
packages we're looking for.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>