tests: conditionally install `moby-tini`
Until the "default" package installed by the repo is the new
`moby-engine` *without* `docker-init` packaged in, `moby-tini` will
conflict with packages other than itself and `moby-engine`.

This is because we used to build `moby-engine` with `docker-init` as
part of that package. The new setup is to leave that out of
`moby-engine`, and put it in its own package `moby-tini`. When the old
`moby-engine` package is installed alongside `moby-tini`, there's a
conflict because they both want to install the file
`/usr/bin/docker-init`.

This workaround will be necessary to solve that problem until the new
`moby-engine` (without `docker-init`) becomes the default.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>