queue: fix error handling
There was a subtle bug in the error handling closure. The `err` variable
from the environment was being captured and appended to the errs array.
That variable was always nil, so no errors were being captured.

There was also a problem where the pipeline was improperly surfacing
errors. This is because the payload is now base64 encoded. The surfacing
code has been adjusted accordingly.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>