pipeline: don't publish tini when not building tini
There was a bug in the code. Because tini isn't in the package repos
yet, we're installing it for tests (except when we're testing tini
itself). This ends up getting published to the queue with a URI
like `https://moby.blob.core.windows.net/1691424061-moby-trigger/moby-engine/24.0.5+azure/mariner2/linux_amd64/moby-tini-0.19.0-1.cm2.x86_64.rpm`.

This is leading to conflicts on the queue consumer side, since there are
multiple versions of tini enqueued at the same time.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>