build: Fix output directory structure
The latest api determines the listed architecture by performing a series
of transformations. For example, `linux_amd64` would be split by the `_`
delimiter, with `"linux"` as the os, `"amd64"` as the architecture, and
`null` as the variant. The variant is `null` because there is no third
element when splitting by `_`.

However, when the directory containing the artifact is named
`linux_armv7` (as the code prior to this commit did), then the variant
will still be `null`, whereas we want the variant to be `"v7"`. The
variant is used by the latest API to assemble the architecture listing.

Existing listings have been `"arm/v7"`, and this change will return
consistency to the listed architecture of armv7 packages.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>