Add support for building containerd 2.0
The main difference here is 2.0 drops support for the v1 shim
(`containerd-shim` and `containerd-shim-runc-v1`) binaries.
But we still need to support building 1.x packages for now so this
splits those up into "archives".
When building we do a switch on the version tag to determine which set
of archives to use.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>