Fix go version switch for containerd
I tested this using a different version format than what we actually use
in our builds.
Since the tag field is used for package versions, the version is not
semver compatible due to the use of `~` for prerelease versions.

This just drops the tilde and everything after since this is not
something we care about when checking the version.