runc: fix how version is set after upstream change
Upstream changed how the binary version can be set which broke how we
were setting it.
Now we can pass a VERSION env var into runcs make file to have it use
that instead of modifying the VERSION file in the repo.

Because we already have a VERSION field set in the env, the runc
makefile is taking that, but it does not contain the revision number, so
we need to modify that env.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>