runc: set COMMIT via make arg
The makefile for runc no longer supports setting COMMIT (and some other
things) through env var and instead must be set through make args.

Additionally adds `urfave_cli_no_docs` which is added to the base
BUILDTAGS in 1.2 but is not set in 1.1.
This just makes it so the build tag is set on both 1.1 and 1.2 instead
of relying on the make file for those tags across branches.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>